GROM-140: Link session DB to rocky request call

+++ app/src/main/java/com/rockthevote/grommet/ui/registration/SessionData.kt
@@ -8,7 +8,7 @@ import com.rockthevote.grommet.data.db.model.GeoLocation
  *  database entry
  */
 data class SessionData(
-    val partnerId: Int,
+    val partnerId: Long,
     val canvasserName: String,
     val sourceTrackingId: String,
     val partnerTrackingId: String,
